[Dream3D][7][3D Mesh-Batching]

+++ src/Engine/TGA.cpp
@@ -58,7 +58,7 @@ int TGAImg::Load(char* szFilename)
 
 	// Get file size
 	fIn.seekg(0,ios_base::end);
-	ulSize=fIn.tellg();
+	ulSize = (unsigned long)fIn.tellg();
 	fIn.seekg(0,ios_base::beg);
 
 	// Allocate some space
